homestud/requirements.txt
Package           Version
----------------- ---------
asgiref           3.2.10
certifi           2020.6.20
chardet           3.0.4
defusedxml        0.6.0
Django            3.0.8
django-allauth    0.42.0
django-formtools  2.2
idna              2.10
numpy             1.19.0
oauthlib          3.1.0
Pillow            7.2.0
pip               20.1.1
psycopg2          2.8.5
python3-openid    3.2.0
pytz              2020.1
requests          2.24.0
requests-oauthlib 1.3.0
setuptools        46.4.0
sqlparse          0.3.1
urllib3           1.25.9
wheel             0.34.2
